+++ scripts/remove-gnu-softwares.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
 
-if  [[ $(command brew ls | grep gnu-sed) != "" ]]
+brew --version > /dev/null 2>&1
+
+if [ $? -eq 0 ]
 then
-  printf "Removing gnu-sed (via Homebrew)..\n"
-  brew remove --force gnu-sed
-  printf "gnu-sed (via Homebrew) was removed!\n\n"
+  if  [[ $(command brew ls | grep gnu-sed) != "" ]]
+  then
+    printf "Removing gnu-sed (via Homebrew)..\n"
+    brew remove --force gnu-sed
+    printf "gnu-sed (via Homebrew) was removed!\n\n"
+  else
+    printf "Looks like gnu-sed (via Homebrew) is not installed.. Everything good here!\n\n"
+  fi
 else
-  printf "Looks like gnu-sed (via Homebrew) is not installed.. Everything good here!\n\n"
+  printf "Brew is necessary here.... skipping...\n\n"
 fi
